+++ tests/integration/cases/dict_mixed_int_widths/Go.go
@@ -1,0 +1,10 @@
+package main
+
+func main() {
+my_data := map[string]any{
+	"a": 1,
+	"b": 3000000000,
+	"c": "x",
+}
+_ = my_data
+}