+++ tests/integration/cases/dict_mixed_int_widths/ObjectiveC.m
@@ -1,0 +1,9 @@
+#import <Foundation/Foundation.h>
+void check_(void) {
+id my_data = @{
+    @"a": @(1),
+    @"b": @(3000000000),
+    @"c": @"x",
+};
+    (void)my_data;
+}